# docs/assets/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IBEF Backend API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/sensor/{sensor_id}/data": {
      "get": {
        "tags": [
          "sensor"
        ],
        "summary": "Get Sensor Data",
        "description": "Get the latest data point from a sensor (calibrated/processed value).\nTime is relative to test start if a test is running, otherwise 0.",
        "operationId": "get_sensor_data_api_sensor__sensor_id__data_get",
        "parameters": [
          {
            "name": "sensor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sensor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Point"
                }
              }
            }
          },
          "400": {
            "description": "Invalid sensor_id provided.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Invalid sensor_id: INVALID. Valid values are: FORCE, DISP_1, DISP_2, DISP_3"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sensor/{sensor_id}/data/history": {
      "get": {
        "tags": [
          "sensor"
        ],
        "summary": "Get Sensor Data History",
        "description": "Get historical data points from a sensor (calibrated/processed values).\nWindow is expressed in seconds and must be one of: 30, 60, 120, 300, 600.\nReturns a fixed number of points (based on 30s sampling) evenly spaced across the window.",
        "operationId": "get_sensor_data_history_api_sensor__sensor_id__data_history_get",
        "parameters": [
          {
            "name": "sensor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sensor Id"
            }
          },
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30,
              "title": "Window"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PointsList"
                }
              }
            }
          },
          "400": {
            "description": "Invalid sensor_id or window parameter.",
            "content": {
              "application/json": {
                "examples": {
                  "invalid_sensor": {
                    "value": {
                      "detail": "Invalid sensor_id: INVALID. Valid values are: FORCE, DISP_1, DISP_2, DISP_3"
                    }
                  },
                  "invalid_window": {
                    "value": {
                      "detail": "Invalid window: 45. Allowed values are: [30, 60, 120, 300, 600]"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "No test is currently running.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "No test is currently running"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sensor/{sensor_id}/raw": {
      "get": {
        "tags": [
          "sensor"
        ],
        "summary": "Get Sensor Raw Data",
        "description": "Get the latest raw (uncalibrated) data point from a sensor.\nTime is relative to test start if a test is running, otherwise 0.",
        "operationId": "get_sensor_raw_data_api_sensor__sensor_id__raw_get",
        "parameters": [
          {
            "name": "sensor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sensor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Point"
                }
              }
            }
          },
          "400": {
            "description": "Invalid sensor_id provided.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Invalid sensor_id: INVALID. Valid values are: FORCE, DISP_1, DISP_2, DISP_3"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sensor/{sensor_id}/zero": {
      "get": {
        "tags": [
          "sensor"
        ],
        "summary": "Get Sensor Zero Offset",
        "description": "Get the current zero offset applied to raw readings for a sensor.\nThe offset represents how much raw values are shifted to compute calibrated values.",
        "operationId": "get_sensor_zero_offset_api_sensor__sensor_id__zero_get",
        "parameters": [
          {
            "name": "sensor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sensor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OffsetResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid sensor_id provided.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Invalid sensor_id: INVALID. Valid values are: FORCE, DISP_1, DISP_2, DISP_3"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "sensor"
        ],
        "summary": "Zero Sensor",
        "description": "Zero a sensor by recording its current value.\nFuture readings from this sensor will be adjusted by subtracting this value.",
        "operationId": "zero_sensor_api_sensor__sensor_id__zero_put",
        "parameters": [
          {
            "name": "sensor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sensor Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid sensor_id provided.",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Invalid sensor_id: INVALID. Valid values are: FORCE, DISP_1, DISP_2, DISP_3"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OffsetResponse": {
        "properties": {
          "offset": {
            "type": "number",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "offset"
        ],
        "title": "OffsetResponse"
      },
      "Point": {
        "properties": {
          "time": {
            "type": "number",
            "title": "Time"
          },
          "value": {
            "type": "number",
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "time",
          "value"
        ],
        "title": "Point"
      },
      "PointsList": {
        "properties": {
          "list": {
            "items": {
              "$ref": "#/components/schemas/Point"
            },
            "type": "array",
            "title": "List"
          }
        },
        "type": "object",
        "required": [
          "list"
        ],
        "title": "PointsList"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}